+++ src/app/(main)/chat/(workspace)/_layout/Desktop/ChatHeader/index.tsx
@@ -1,0 +1,8 @@
+import { ChatHeader } from '@lobehub/ui';
+
+import HeaderAction from './HeaderAction';
+import Main from './Main';
+
+const Header = () => <ChatHeader left={<Main />} right={<HeaderAction />} style={{ zIndex: 11 }} />;
+
+export default Header;