+++ src/styles/global.ts
@@ -1,22 +1,25 @@
 import { Theme, css } from 'antd-style';
 
-export default ({ prefixCls, token }: { prefixCls: string; token: Theme }) => css`
+// fix ios input keyboard
+// overflow: hidden;
+// ref: https://zhuanlan.zhihu.com/p/113855026
+export default ({ token }: { prefixCls: string; token: Theme }) => css`
   html,
   body,
-  #__next,
-  .${prefixCls}-app {
+  #__next {
     position: relative;
+
     overscroll-behavior: none;
-    height: 100% !important;
-    max-height: 100dvh !important;
+
+    height: 100%;
+    min-height: 100dvh;
+    max-height: 100dvh;
+
+    background: ${token.colorBgLayout};
   }
 
   * {
     scrollbar-color: ${token.colorFill} transparent;
     scrollbar-width: thin;
-  }
-
-  p {
-    margin-bottom: 0;
   }
 `;